Merged changes from OasisComplexModel master

+++ complex_model/RiskFrontiers_HailAUS_gulcalc.py
@@ -11,15 +11,12 @@ import complex_model.DefaultSettings as DS
 from backports.tempfile import TemporaryDirectory
 from oasislmf.utils.exceptions import OasisException
 from complex_model.OasisToRF import create_rf_input, DEFAULT_DB, get_connection_string, is_valid_model_data
-from complex_model.GulcalcToBin import gulcalc_sqlite_to_bin
+from complex_model.GulcalcToBin import gulcalc_sqlite_fp_to_bin
 from complex_model.Common import PerilSet
+from complex_model.RFException import FileNotFoundException
 from datetime import datetime
 import multiprocessing
 
-
-_DEBUG = False
-
-logging.basicConfig(stream=sys.stdout, level=logging.DEBUG)
 
 PY3K = sys.version_info >= (3, 0)
 
@@ -35,11 +32,17 @@ else:
         msvcrt.setmode(sys.stdout.fileno(), os.O_BINARY)
     output_stdout = sys.stdout
 
+_DEBUG = False
+_WORKER_LOG_FILE = "/var/log/oasis/worker.log"
+logging.basicConfig(level=logging.DEBUG if _DEBUG else logging.INFO,
+                    filename=_WORKER_LOG_FILE,
+                    format='[%(asctime)s: %(levelname)s/%(filename)s] %(message)s')
+
 
 def clean_directory(dir_path):
     import shutil
-    shutil.rmtree(dir_path)
-    import os
+    if os.path.exists(dir_path) and os.path.isdir(dir_path):
+        shutil.rmtree(dir_path)
     os.mkdir(dir_path)
 
 
@@ -68,6 +71,14 @@ def main():
     parser.add_argument(
         '-c', '--coverage_output_stream', required=False, default=None,
         help='Coverage output stream.',
+    )
+    parser.add_argument(
+        '-M', '--model_data_directory', required=False, default=DS.MODEL_DATA_DIRECTORY,
+        help='Model data directory.',
+    )
+    parser.add_argument(
+        '-X', '--complex_model_directory', required=False, default=DS.COMPLEX_MODEL_DIRECTORY,
+        help='Complex model directory.',
     )
 
     args = parser.parse_args()
@@ -103,9 +114,8 @@ def main():
     if not os.path.exists(inputs_fp):
         raise Exception('Inputs directory does not exist')
 
-    inputs_fp_csv = os.path.join(inputs_fp, "csv")
     complex_items_filename = "complex_items.csv"  # args.complex_items_filename
-    complex_items_fp = os.path.join(inputs_fp_csv, complex_items_filename)
+    complex_items_fp = os.path.join(inputs_fp, complex_items_filename)
     if not os.path.exists(complex_items_fp):
         raise Exception('Complex items file does not exist')
 
@@ -120,7 +130,7 @@ def main():
         model_settings = {}
 
     # Read the inputs, including the extended items
-    with open(os.path.join(inputs_fp_csv, 'coverages.csv')) as p:
+    with open(os.path.join(inputs_fp, 'coverages.csv')) as p:
         coverages_pd = pd.read_csv(p)
 
     # with open(os.path.join(inputs_fp, 'gulsummaryxref.csv')) as p:
@@ -128,24 +138,44 @@ def main():
 
     with open(complex_items_fp) as p:
         items_pd = pd.read_csv(p)
+
     with TemporaryDirectory() as working_dir:
         log_filename = "worker_{}_{}.log".format(event_batch, datetime.now().strftime("%Y%m%d%H%M%S"))
-        log_file = os.path.join(DS.WORKER_LOG_DIRECTORY, log_filename)
+        # log_filename = "worker.log"
+        log_fp = os.path.join(DS.WORKER_LOG_DIRECTORY, log_filename)
         if _DEBUG:
-            working_dir = "/hadoop/oasis/tmp"
+            working_dir = "/tmp/oasis_debug_{}".format(event_batch)
             clean_directory(working_dir)
-            log_file = os.path.join(working_dir, log_filename)
+            log_fp = os.path.join(working_dir, log_filename)
+        logging.info("Working directory for worker with batch " + str(event_batch) + " is set to be " + working_dir)
+        logging.info("The process independent log file for this worker is " + log_filename)
+
         # Write out RF canonical input files
-        risk_platform_data = os.path.join(DS.MODEL_DATA_DIRECTORY)
+        risk_platform_data = os.path.join(args.model_data_directory)
         if not is_valid_model_data(risk_platform_data):
-            raise FileNotFoundError("Model data not set correctly: " + risk_platform_data)
+            message = "Model data not set correctly: " + risk_platform_data
+            logging.error(message)
+            raise FileNotFoundException(message, 420)
         temp_db_fp = os.path.join(working_dir, DEFAULT_DB)
+        logging.info("RF model data file found in " + risk_platform_data)
+
+        # check RF license exists
+        licence_file = os.path.join(risk_platform_data, "license.txt")
+        if not os.path.isfile(licence_file):
+            licence_file = os.path.join(risk_platform_data, "licence.txt")
+            if not os.path.isfile(licence_file):
+                message = "License file not found at " + risk_platform_data
+                logging.error(message)
+                raise FileNotFoundException(message, 410)
+        logging.info("License file found at " + licence_file)
 
         # populate RF exposure and coverage datatable
+        logging.info("STARTED: Generating RF input database in " + temp_db_fp)
         num_rows = create_rf_input(items_pd, coverages_pd, temp_db_fp, risk_platform_data)
+        logging.info("COMPLETED: RF input database generated in " + temp_db_fp + " [OK]")
 
         # generate oasis_param.json
-        complex_model_directory = DS.COMPLEX_MODEL_DIRECTORY
+        complex_model_directory = args.complex_model_directory
         max_event_id = PerilSet[model_version_id]['MAX_EVENT_INDEX']
         num_cores = multiprocessing.cpu_count()
         max_parallelism = int(max(1, min(num_cores, num_cores/max_event_batch)))
@@ -159,7 +189,7 @@ def main():
             "NumSamples": int(number_of_samples),
             "CountryCode": DS.COUNTRY_CODE,
             "ComplexModelDirectory": complex_model_directory,
-            "LicenseFile": os.path.join(risk_platform_data, "license.txt"),
+            "LicenseFile": licence_file,
             "RiskPlatformData": risk_platform_data,
             "WorkingDirectory": working_dir,
             "NumRows": num_rows,
@@ -169,29 +199,45 @@ def main():
             if 'individual_risk_mode' in model_settings else DS.DEFAULT_INDIVIDUAL_RISK_MODE,
             "StaticMotor": model_settings['static_motor']
             if 'static_motor' in model_settings else DS.DEFAULT_STATIC_MOTOR,
+            "DemandSurge": model_settings['demand_surge']
+            if 'demand_surge' in model_settings else DS.DEFAULT_DEMAND_SURGE,
+            "InputScaling": model_settings['input_scaling']
+            if 'input_scaling' in model_settings else DS.DEFAULT_INPUT_SCALING,
         }
 
         oasis_param_fp = os.path.join(working_dir, "oasis_param.json")
         with open(oasis_param_fp, 'w') as param:
             param.writelines(json.dumps(oasis_param, indent=4, separators=(',', ': ')))
+            logging.debug("The Risk Frontiers .Net engine will be called with the following parameters")
+            logging.debug(json.dumps(oasis_param, indent=4, separators=(',', ':')))
 
         # call Risk.Platform.Core/Risk.Platform.Core.dll --oasis -c oasis_param.json
         cmd_str = "{} --oasis -c {} {} --log {}".format(os.path.join(oasis_param["ComplexModelDirectory"],
                                                                      "Risk.Platform.Core", "Risk.Platform.Core"),
                                                         oasis_param_fp,
                                                         "--debug" if _DEBUG else "",
-                                                        log_file)
+                                                        log_fp)
         try:
+            logging.info("STARTED: Calling Risk Frontiers .Net engine: " + cmd_str + " for event batch "
+                         + str(event_batch))
             subprocess.check_call(cmd_str, stderr=subprocess.STDOUT, shell=True)
+            logging.info("COMPLETED: Loss database has been generated in " + temp_db_fp + " for event batch "
+                         + str(event_batch))
             if do_item_output:
-                gulcalc_sqlite_to_bin(temp_db_fp, output_item, int(number_of_samples), 1)
+                logging.info("STARTED: Transforming sqlite losses into gulcalc item binary stream for event_batch "
+                             + str(event_batch))
+                gulcalc_sqlite_fp_to_bin(temp_db_fp, output_item, int(number_of_samples), 1)
             if do_coverage_output:
-                gulcalc_sqlite_to_bin(temp_db_fp, output_coverage, int(number_of_samples), 2)
-
+                logging.info("STARTED: Transforming sqlite losses into gulcalc coverage binary stream for event_batch "
+                             + str(event_batch))
+                gulcalc_sqlite_fp_to_bin(temp_db_fp, output_coverage, int(number_of_samples), 2)
+            logging.info("COMPLETED: Successfully generated losses as gulcalc binary stream for event batch "
+                         + str(event_batch))
         except subprocess.CalledProcessError as e:
+            logging.error("An error occurred while calling the Risk Frontiers .Net engine. Please look at " + log_fp +
+                          " for more information regarding this issue")
             raise OasisException(e)
 
 
 if __name__ == "__main__":
     main()
-
